% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tsa_cor}
\alias{print.tsa_cor}
\title{Print a tsa_cor object}
\usage{
\method{print}{tsa_cor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tsa_cor"}.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{"tsa_cor"}, namely the same object
  \code{x} supplied to the method, returned invisibly. The method
  prints a concise summary of the Trial Sequential Analysis results
  and returns the original object unchanged.
}
\description{
Print a tsa_cor object
}
