% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tsacor-package}
\alias{tsacor}
\alias{tsacor-package}
\title{tsacor: Trial Sequential Analysis for Meta-Analyses of Correlations}
\description{
Performs Trial Sequential Analysis (TSA) for meta-analyses of Pearson product-moment correlations and Spearman rank correlations pooled on Fisher's z scale. Study-level correlations are transformed with atanh(), their standard errors are taken from the reported confidence intervals or from the sample sizes (with the Fieller-Hartley-Pearson or Bonett-Wright variance models for Spearman's rho), and the required information size is corrected for heterogeneity with the Diversity (D-squared) adjustment of Wetterslev et al. (2009). O'Brien-Fleming-type alpha- and beta-spending trial sequential monitoring boundaries are computed at the inverse-variance information based on the observed study-level standard errors of Fisher's z, via a compiled (C++) recursive numerical integration engine ported from the R package 'RTSA' (Soerensen, Olsen, Lange and Gluud; an R implementation of the Trial Sequential Analysis software of the Copenhagen Trial Unit, \url{https://ctu.dk/tools}), with no fixed software-imposed limit on the number of looks (subject to available computational resources). Produces a cumulative Z-curve plot against the cumulative number of participants with efficacy, futility, and conventional significance boundaries. The engine is shared with the sister package 'tsahr' for hazard ratios. Methodology follows Wetterslev et al. (2009) \doi{10.1186/1471-2288-9-86}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tarak-dhaouadi/tsacor}
  \item Report bugs at \url{https://github.com/tarak-dhaouadi/tsacor/issues}
}

}
\author{
\strong{Maintainer}: Tarak Dhaouadi \email{dhaouaditarak@yahoo.fr}

Authors:
\itemize{
  \item Tarak Dhaouadi \email{dhaouaditarak@yahoo.fr}
}

Other contributors:
\itemize{
  \item Anne Lyngholm Soerensen (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Markus Harboe Olsen (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Theis Lange (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Christian Gluud (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
}

}
